api-code/requirements.txt
Flask==1.1.2
geobuf==1.1.1
geopandas==0.8.1
gunicorn==20.0.4
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
plotly==4.14.3
requests==2.25.1
pyarrow
Shapely==1.7.1
Pillow==7.1.1
